Add include/excludeVersions to BranchSpec
This acts as an addition/subtraction combo that is only used if the versions field is not set. Useful for April Fools snapshots.
Add settings.gradle for Gradle JDK resolution. Fixes case when user does not have the necessary JDK downloaded (like me not having J21 for 24w14a) and tries to import with Gradle.
Fix bug calculating target version if it is omitted
Change initial commit generation to only copy gradle/wrapper instead of gradle to account for a potential future use of libs.versions.toml by excluding it
Fix parsing of --cfg parameter when provided URI uses the file scheme
Update Gradle to 8.7