Use URIish for Git remote instead of URL
The URL class checks whether there is a stream factory for the given
protocol of the URL; in other words, it checks if it recognizes the URL
protocol.

Unfortunately, the `ssh` protocol, commonly used in Git remotes, is not
recognized. This means trying to construct an SSH URL fails with an
exception.

To fix this, we use URIish, which implements no such checks. URIish
belongs to JGit, and is used instead of URI as jopt-simple is able to
handle the URIish constructor, thus avoiding manually converting a URI
to a URIish for use by JGit.